refactor: 重构 UploadService 到 Actor 模式

+++ feature/gallery/core/src/main/kotlin/plutoproject/feature/gallery/core/decode/DecodeConstraints.kt
@@ -1,7 +1,7 @@
 package plutoproject.feature.gallery.core.decode
 
 class DecodeConstraints(
-    val maxBytes: Int,
+    val maxBytes: Long,
     val maxPixels: Int,
     val maxFrames: Int,
 ) {
